Update stats in API when game is over

+++ Unity/Assets/Scripts/Models/GameResults.cs
@@ -2,7 +2,7 @@ namespace LD51.Unity.Models
 {
     public class GameResults
     {
-        public ushort Score { get; set; }
-        public ushort Waves { get; set; }
+        public int Score { get; set; }
+        public int Waves { get; set; }
     }
 }
