+++ Unity/Assets/Scripts/Models/Player.cs
@@ -9,8 +9,8 @@ namespace LD51.Unity.Models
         public string DisplayName { get; set; }
         public string ShirtColor { get; set; }
         public string PantColor { get; set; }
-        public ushort Score { get; set; }
-        public ushort Waves { get; set; }
+        public int Score { get; set; }
+        public int Waves { get; set; }
 
         public Player()
         {
